+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "note-organizer" extension will be documented in this
 
 ## [Unreleased]
  - Project and note scanning: find all of the project instead of only those with notes.
+ - Toggle to hide and show empty project in tree data.
 
 ## [0.2.0]